# added logging to help diagnose performance