--- a/crates/nu_plugin_inc/src/main.rs
+++ b/crates/nu_plugin_inc/src/main.rs
@@ -1,6 +1,11 @@
 use nu_plugin::serve_plugin;
 use nu_plugin_inc::Inc;
 
+use log::{info, trace, warn};
+use simple_logger::SimpleLogger;
+
 fn main() {
+    SimpleLogger::new().init().unwrap();
+    trace!("main main");
     serve_plugin(&mut Inc::new())
 }