this isn't a real PR - just added logging so if someone wants to try it, they don't have to type all this mess that i did.

![image](https://user-images.githubusercontent.com/343840/140563140-9a685451-4653-4bd1-9eef-e6821b804dc0.png)