@fdncred 2021-11-08
i wrote a script to benchmark the inc plugin and then analyze the results using this PR with logging and this command `benchmark { 1..100 | each { $it | inc }}`

[perf.nu script](https://github.com/nushell/engine-q/files/7497674/perf.txt)
[old perf log](https://github.com/nushell/engine-q/files/7497624/100.txt)
[new perf log](https://github.com/nushell/engine-q/files/7497625/101.txt)

old performance before latest changes
```
╭────┬──────┬──────╮
│ #  │ row  │ perf │
├────┼──────┼──────┤
│  0 │   12 │ 49ms │
│  1 │   24 │ 51ms │
│  2 │   36 │ 46ms │
│  3 │   48 │ 50ms │
│  4 │   60 │ 49ms │
│  5 │   72 │ 50ms │
│  6 │   84 │ 50ms │
│  7 │   96 │ 53ms │
│  8 │  108 │ 47ms │
│  9 │  120 │ 47ms │
│ 10 │  132 │ 48ms │
│ 11 │  144 │ 45ms │
│ 12 │  156 │ 43ms │
│ 13 │  168 │ 45ms │
│ 14 │  180 │ 44ms │
│ 15 │  192 │ 46ms │
│ 16 │  204 │ 45ms │
│ 17 │  216 │ 45ms │
│ 18 │  228 │ 50ms │
│ 19 │  240 │ 47ms │
│ 20 │  252 │ 44ms │
│ 21 │  264 │ 41ms │
│ 22 │  276 │ 48ms │
│ 23 │  288 │ 48ms │
│ 24 │  300 │ 42ms │
│ 25 │  312 │ 44ms │
│ 26 │  324 │ 48ms │
│ 27 │  336 │ 49ms │
│ 28 │  348 │ 49ms │
│ 29 │  360 │ 51ms │
│ 30 │  372 │ 49ms │
│ 31 │  384 │ 53ms │
│ 32 │  396 │ 46ms │
│ 33 │  408 │ 48ms │
│ 34 │  420 │ 46ms │
│ 35 │  432 │ 68ms │
│ 36 │  444 │ 56ms │
│ 37 │  456 │ 51ms │
│ 38 │  468 │ 52ms │
│ 39 │  480 │ 52ms │
│ 40 │  492 │ 50ms │
│ 41 │  504 │ 52ms │
│ 42 │  516 │ 52ms │
│ 43 │  528 │ 48ms │
│ 44 │  540 │ 56ms │
│ 45 │  552 │ 52ms │
│ 46 │  564 │ 53ms │
│ 47 │  576 │ 54ms │
│ 48 │  588 │ 55ms │
│ 49 │  600 │ 55ms │
│ 50 │  612 │ 59ms │
│ 51 │  624 │ 57ms │
│ 52 │  636 │ 60ms │
│ 53 │  648 │ 51ms │
│ 54 │  660 │ 54ms │
│ 55 │  672 │ 58ms │
│ 56 │  684 │ 52ms │
│ 57 │  696 │ 51ms │
│ 58 │  708 │ 50ms │
│ 59 │  720 │ 54ms │
│ 60 │  732 │ 54ms │
│ 61 │  744 │ 52ms │
│ 62 │  756 │ 54ms │
│ 63 │  768 │ 57ms │
│ 64 │  780 │ 54ms │
│ 65 │  792 │ 58ms │
│ 66 │  804 │ 57ms │
│ 67 │  816 │ 55ms │
│ 68 │  828 │ 58ms │
│ 69 │  840 │ 65ms │
│ 70 │  852 │ 61ms │
│ 71 │  864 │ 56ms │
│ 72 │  876 │ 54ms │
│ 73 │  888 │ 52ms │
│ 74 │  900 │ 53ms │
│ 75 │  912 │ 50ms │
│ 76 │  924 │ 53ms │
│ 77 │  936 │ 54ms │
│ 78 │  948 │ 52ms │
│ 79 │  960 │ 53ms │
│ 80 │  972 │ 57ms │
│ 81 │  984 │ 55ms │
│ 82 │  996 │ 55ms │
│ 83 │ 1008 │ 55ms │
│ 84 │ 1020 │ 57ms │
│ 85 │ 1032 │ 59ms │
│ 86 │ 1044 │ 61ms │
│ 87 │ 1056 │ 60ms │
│ 88 │ 1068 │ 66ms │
│ 89 │ 1080 │ 66ms │
│ 90 │ 1092 │ 58ms │
│ 91 │ 1104 │ 52ms │
│ 92 │ 1116 │ 52ms │
│ 93 │ 1128 │ 49ms │
│ 94 │ 1140 │ 50ms │
│ 95 │ 1152 │ 51ms │
│ 96 │ 1164 │ 50ms │
│ 97 │ 1176 │ 55ms │
│ 98 │ 1188 │ 57ms │
╰────┴──────┴──────╯
```
new performance after latest changes
```
╭────┬──────┬──────╮
│ #  │ row  │ perf │
├────┼──────┼──────┤
│  0 │   12 │ 55ms │
│  1 │   24 │ 37ms │
│  2 │   36 │ 40ms │
│  3 │   48 │ 43ms │
│  4 │   60 │ 48ms │
│  5 │   72 │ 46ms │
│  6 │   84 │ 42ms │
│  7 │   96 │ 44ms │
│  8 │  108 │ 42ms │
│  9 │  120 │ 42ms │
│ 10 │  132 │ 45ms │
│ 11 │  144 │ 42ms │
│ 12 │  156 │ 41ms │
│ 13 │  168 │ 46ms │
│ 14 │  180 │ 41ms │
│ 15 │  192 │ 40ms │
│ 16 │  204 │ 45ms │
│ 17 │  216 │ 44ms │
│ 18 │  228 │ 39ms │
│ 19 │  240 │ 49ms │
│ 20 │  252 │ 43ms │
│ 21 │  264 │ 39ms │
│ 22 │  276 │ 38ms │
│ 23 │  288 │ 46ms │
│ 24 │  300 │ 43ms │
│ 25 │  312 │ 44ms │
│ 26 │  324 │ 40ms │
│ 27 │  336 │ 44ms │
│ 28 │  348 │ 46ms │
│ 29 │  360 │ 49ms │
│ 30 │  372 │ 42ms │
│ 31 │  384 │ 43ms │
│ 32 │  396 │ 42ms │
│ 33 │  408 │ 44ms │
│ 34 │  420 │ 44ms │
│ 35 │  432 │ 41ms │
│ 36 │  444 │ 43ms │
│ 37 │  456 │ 42ms │
│ 38 │  468 │ 37ms │
│ 39 │  480 │ 41ms │
│ 40 │  492 │ 50ms │
│ 41 │  504 │ 53ms │
│ 42 │  516 │ 49ms │
│ 43 │  528 │ 51ms │
│ 44 │  540 │ 51ms │
│ 45 │  552 │ 56ms │
│ 46 │  564 │ 52ms │
│ 47 │  576 │ 48ms │
│ 48 │  588 │ 52ms │
│ 49 │  600 │ 56ms │
│ 50 │  612 │ 51ms │
│ 51 │  624 │ 53ms │
│ 52 │  636 │ 50ms │
│ 53 │  648 │ 48ms │
│ 54 │  660 │ 51ms │
│ 55 │  672 │ 53ms │
│ 56 │  684 │ 47ms │
│ 57 │  696 │ 48ms │
│ 58 │  708 │ 49ms │
│ 59 │  720 │ 49ms │
│ 60 │  732 │ 52ms │
│ 88 │ 1068 │ 50ms │
│ 89 │ 1080 │ 45ms │
│ 90 │ 1092 │ 53ms │
│ 91 │ 1104 │ 51ms │
│ 92 │ 1116 │ 52ms │
│ 93 │ 1128 │ 53ms │
│ 94 │ 1140 │ 51ms │
│ 95 │ 1152 │ 56ms │
│ 96 │ 1164 │ 44ms │
│ 97 │ 1176 │ 50ms │
│ 98 │ 1188 │ 46ms │
╰────┴──────┴──────╯
```