@fdncred 2021-11-20
@elferherrera do you know if `register` still runs twice like it shows in the logging? it seems like you and jt were talking about that recently.