@elferherrera 2021-11-22
@fdncred yeap. It runs twice, but it doesnt register anything until you press enter. In theory, every command is parsed every loop. It seems that the extra run is done by the linter